refactor - 도메인패키지추가.
