+++ src/main/java/lotto/domain/Lotto.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.domain;
+
+import lotto.domain.LottoNumbers;
 
 import java.util.ArrayList;
 import java.util.List;
