+++ src/test/java/lotto/LottoPriceTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoPrice;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
